add tools for test persist with bot
