+++ src/main/java/com/community/tools/controller/GitHubController.java
@@ -63,71 +63,4 @@ public class GitHubController {
     List<EventData> eventData = gitHubService.getEvents(start, end);
     return ok().body(eventData);
   }
-
-  @GetMapping(value = "/sendTestMessage", produces = MediaType.APPLICATION_JSON_VALUE)
-  public void getAllEvents() throws ParseException {
-    String message = "[\n"
-        + "\t{\n"
-        + "\t\t\"type\": \"image\",\n"
-        + "\t\t\"title\": {\n"
-        + "\t\t\t\"type\": \"plain_text\",\n"
-        + "\t\t\t\"text\": \"image1\",\n"
-        + "\t\t\t\"emoji\": true\n"
-        + "\t\t},\n"
-        + "\t\t\"image_url\": \"https://api.slack.com/img/blocks/bkb_template_images/beagle.png\",\n"
-        + "\t\t\"alt_text\": \"image1\"\n"
-        + "\t},\n"
-        + "\t{\n"
-        + "\t\t\"type\": \"section\",\n"
-        + "\t\t\"text\": {\n"
-        + "\t\t\t\"type\": \"mrkdwn\",\n"
-        + "\t\t\t\"text\": \"Read and confirm that you agree to our <https://www.youtube.com/watch?v=O6YzU00oack|rules> BOY :v:. \"\n"
-        + "\t\t},\n"
-        + "\t\t\"accessory\": {\n"
-        + "\t\t\t\"type\": \"button\",\n"
-        + "\t\t\t\"text\": {\n"
-        + "\t\t\t\t\"type\": \"plain_text\",\n"
-        + "\t\t\t\t\"text\": \"Agree\",\n"
-        + "\t\t\t\t\"emoji\": true\n"
-        + "\t\t\t},\n"
-        + "\t\t\t\"value\": \"click_me_123\"\n"
-        + "\t\t}\n"
-        + "\t}\n"
-        + "]";
-    try {
-      slackService.sendEventsMessage("roman", message);
-    } catch (IOException | SlackApiException e) {
-      e.printStackTrace();
-    }
-  }
-
-  @GetMapping(value = "/sendTestMessage2", produces = MediaType.APPLICATION_JSON_VALUE)
-  public void getSendMessage() throws ParseException {
-    String message = "[\n"
-        + "\t{\n"
-        + "\t\t\"type\": \"divider\"\n"
-        + "\t},\n"
-        + "\t{\n"
-        + "\t\t\"type\": \"section\",\n"
-        + "\t\t\"text\": {\n"
-        + "\t\t\t\"type\": \"mrkdwn\",\n"
-        + "\t\t\t\"text\": \"This is not work button. \"\n"
-        + "\t\t},\n"
-        + "\t\t\"accessory\": {\n"
-        + "\t\t\t\"type\": \"button\",\n"
-        + "\t\t\t\"text\": {\n"
-        + "\t\t\t\t\"type\": \"plain_text\",\n"
-        + "\t\t\t\t\"text\": \"Button\",\n"
-        + "\t\t\t\t\"emoji\": true\n"
-        + "\t\t\t},\n"
-        + "\t\t\t\"value\": \"click_me_123\"\n"
-        + "\t\t}\n"
-        + "\t}\n"
-        + "]";
-    try {
-      slackService.sendEventsMessage("roman", message);
-    } catch (IOException | SlackApiException e) {
-      e.printStackTrace();
-    }
-  }
 }
